Refactor Achievements into Feature structure
Move all achievement code to Features/Achievements/ following the same pattern as Patreon feature. Updates namespaces and service registration throughout codebase.
