+++ backend/Veloci.Logic/Bot/Discord/DiscordMessageEventHandler.cs
@@ -5,6 +5,7 @@ using Veloci.Data.Domain;
 using Veloci.Data.Repositories;
 using Veloci.Logic.Helpers;
 using Veloci.Logic.Notifications;
+using Veloci.Logic.Features.Achievements.Notifications;
 using Veloci.Logic.Services;
 
 namespace Veloci.Logic.Bot.Discord;
