+++ backend/Veloci.Logic/Bot/Telegram/Commands/PilotCommand.cs
@@ -2,7 +2,7 @@ using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
 using Veloci.Data.Domain;
 using Veloci.Data.Repositories;
-using Veloci.Logic.Achievements.Base;
+using Veloci.Logic.Features.Achievements.Base;
 using Veloci.Logic.Bot.Telegram.Commands.Core;
 
 namespace Veloci.Logic.Bot.Telegram.Commands;
