+++ backend/Veloci.Logic/Features/Achievements/Collection/FirstPlaceInRaceAchievement.cs
@@ -1,7 +1,7 @@
 using Veloci.Data.Domain;
-using Veloci.Logic.Achievements.Base;
+using Veloci.Logic.Features.Achievements.Base;
 
-namespace Veloci.Logic.Achievements.Collection;
+namespace Veloci.Logic.Features.Achievements.Collection;
 
 public class FirstPlaceInRaceAchievement : IAchievementAfterCompetition
 {
@@ -12,7 +12,9 @@ public class FirstPlaceInRaceAchievement : IAchievementAfterCompetition
     public async Task<bool> CheckAsync(Pilot pilot, Competition competition)
     {
         if (pilot.HasAchievement(Name))
+        {
             return false;
+        }
 
         return competition.IsPilotAtLocalRank(pilot, 1);
     }
