+++ backend/Veloci.Logic/Features/Achievements/Collection/GlobalFirstPlaceAchievement.cs
@@ -1,22 +1,27 @@
 using Veloci.Data.Domain;
-using Veloci.Logic.Achievements.Base;
+using Veloci.Logic.Features.Achievements.Base;
 
-namespace Veloci.Logic.Achievements.Collection;
+namespace Veloci.Logic.Features.Achievements.Collection;
 
 public class GlobalFirstPlaceAchievement : IAchievementAfterCompetition
 {
     public string Name => "GlobalFirstPlace";
     public string Title => "GOAT";
     public string Description => "Перше місце в загальному лідерборді треку";
+
     public async Task<bool> CheckAsync(Pilot pilot, Competition competition)
     {
         if (pilot.HasAchievement(Name))
+        {
             return false;
+        }
 
         var result = competition.CompetitionResults.GetByPilotName(pilot.Name);
 
         if (result is null)
+        {
             throw new Exception("Result is null. Check the logic");
+        }
 
         return result.GlobalRank == 1;
     }
