+++ backend/Veloci.Logic/Features/Achievements/Collection/LastInRaceAchievement.cs
@@ -1,7 +1,7 @@
 using Veloci.Data.Domain;
-using Veloci.Logic.Achievements.Base;
+using Veloci.Logic.Features.Achievements.Base;
 
-namespace Veloci.Logic.Achievements.Collection;
+namespace Veloci.Logic.Features.Achievements.Collection;
 
 public class LastInRaceAchievement : IAchievementAfterCompetition
 {
@@ -12,7 +12,9 @@ public class LastInRaceAchievement : IAchievementAfterCompetition
     public async Task<bool> CheckAsync(Pilot pilot, Competition competition)
     {
         if (pilot.HasAchievement(Name))
+        {
             return false;
+        }
 
         return competition.GetSlowest()?.PlayerName == pilot.Name;
     }
