+++ backend/Veloci.Logic/Features/Achievements/Collection/SecondPlaceInRaceAchievement.cs
@@ -1,8 +1,8 @@
 using Veloci.Data.Domain;
 using Veloci.Data.Repositories;
-using Veloci.Logic.Achievements.Base;
+using Veloci.Logic.Features.Achievements.Base;
 
-namespace Veloci.Logic.Achievements.Collection;
+namespace Veloci.Logic.Features.Achievements.Collection;
 
 public class SecondPlaceInRaceAchievement : IAchievementAfterCompetition
 {
@@ -16,10 +16,13 @@ public class SecondPlaceInRaceAchievement : IAchievementAfterCompetition
     public string Name => "SecondPlaceInRace";
     public string Title => "Майже лідер";
     public string Description => "Друге місце в гонці";
+
     public async Task<bool> CheckAsync(Pilot pilot, Competition competition)
     {
         if (pilot.HasAchievement(Name))
+        {
             return false;
+        }
 
         return competition.IsPilotAtLocalRank(pilot, 2);
     }
