+++ backend/Veloci.Logic/Features/Achievements/Notifications/DayStreakAchievements.cs
@@ -1,6 +1,6 @@
 using MediatR;
 using Veloci.Data.Domain;
 
-namespace Veloci.Logic.Notifications;
+namespace Veloci.Logic.Features.Achievements.Notifications;
 
 public record DayStreakAchievements(List<Pilot> Pilots) : INotification;
