+++ backend/Veloci.Logic/Features/Achievements/Notifications/GotAchievements.cs
@@ -1,0 +1,6 @@
+using MediatR;
+using Veloci.Logic.Features.Achievements.Services;
+
+namespace Veloci.Logic.Features.Achievements.Notifications;
+
+public record GotAchievements(AchievementCheckResults Results) : INotification;
