--- backend/Veloci.Logic/Notifications/GotAchievements.cs
@@ -1,6 +1,0 @@
-using MediatR;
-using Veloci.Logic.Services.Achievements;
-
-namespace Veloci.Logic.Notifications;
-
-public record GotAchievements(AchievementCheckResults Results) : INotification;
